# frontend/generated/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/admin/users":{"get":{"tags":["admin-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CreateUserResponseDTO"}}}}}}},"post":{"tags":["admin-controller"],"operationId":"createNewUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequestBodyDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateUserResponseDTO"}}}}}}},"/api/admin/users/{id}":{"delete":{"tags":["admin-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}},"patch":{"tags":["admin-controller"],"operationId":"editUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditUserRequestBodyDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EditUserResponseDTO"}}}}}}},"/api/auth/current_user":{"get":{"tags":["authentication-controller"],"operationId":"getCurrentUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CurrentUserResponseDTO"}}}}}}}},"components":{"schemas":{"CreateUserRequestBodyDTO":{"type":"object","properties":{"username":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["password","username"]},"CreateUserResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"role":{"type":"string","minLength":1},"createdAt":{"type":"string","format":"date-time","minLength":1}},"required":["createdAt","id","role","username"]},"EditUserRequestBodyDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}}},"EditUserResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"role":{"type":"string","minLength":1}},"required":["id","role","username"]},"CurrentUserResponseDTO":{"type":"object","properties":{"authenticated":{"type":"boolean"},"user":{"type":"string"}},"required":["user"]}}}}